modules/api-location-create/requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
kafka-python==2.0.2